Add api endpoints for direct channel choosing
Add an API endpoint channel/id to directly switch to channel on this
position in the list of configured channels.
Also add an API endpoint channel/name where data={channel.name} should be
send to directly chose the channel to switch to.
Move the up/down API events to channel/up|down.